bot/requirements.txt
psutil
aiohttp
asyncpg
requests
discord.py
cryptography
python-dotenv